+++ src/trx/game/shell/mod.h
@@ -18,9 +18,12 @@ typedef struct {
 } SHELL_MOD;
 
 void Shell_ScanAvailableMods(void);
+int32_t Shell_GetModCount(void);
+const SHELL_MOD *Shell_GetMod(int32_t index);
 const SHELL_MOD *Shell_GetModByName(const char *name);
 const SHELL_MOD *Shell_GetModByType(
     SHELL_MOD_TYPE mod_type, int32_t engine_version);
+bool Shell_CanSwitchToMod(const SHELL_MOD *mod);
 bool Shell_IsCurrentMod(const char *name);
 
 const char *Shell_GetCommonStringsPath(void);